% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{AdjustSign}
\alias{AdjustSign}
\title{Sign operation on matrix {@code Sigma} according to the sign of {@code AI}}
\usage{
AdjustSign(Sigma, AI)
}
\arguments{
\item{Sigma}{p by p matrix;}

\item{AI}{p by K matrix}
}
\value{
p by p matrix
}
\description{
Sign operation on matrix {@code Sigma} according to the sign of {@code AI}
}
