% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{calArates}
\alias{calArates}
\title{Calculate the matrix sup norm, frobenius norm(scaled) and L_inf norm.}
\usage{
calArates(A)
}
\arguments{
\item{A}{matrix}
}
\value{
vector of three numerical results of c(frob, sup, L_inf).
}
\description{
Calculate the matrix sup norm, frobenius norm(scaled) and L_inf norm.
}
