% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_features.R
\name{calculate_features}
\alias{calculate_features}
\title{Calcuate the LaGrACE features for a given dataset}
\usage{
calculate_features(net_file, input_data, reference_samples, maxCat)
}
\arguments{
\item{net_file}{Path to TXT file containing the reference sample network}

\item{input_data}{Full dataset of with all samples}

\item{reference_samples}{Logical vector indicating which samples belong to
reference group (TRUE) and which do not (FALSE)}

\item{maxCat}{Maximum number of categories for a variable to be treated as discrete}
}
\value{
A list containing a dataframe of the LaGrACE features, a dataframe
indicating which variables were selected for the prediction in calculating each feature
, and a dataframe including continuous/discrete feature information
}
\description{
Calculates the LaGrACE features for a given dataset based on the provided
network learned on a group of reference samples.
}
