% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_features_with_learned_model.R
\name{calculate_features_with_learned_model}
\alias{calculate_features_with_learned_model}
\title{Train LaGrACE model using a given reference dataset and calcuate the LaGrACE features for another dataset}
\usage{
calculate_features_with_learned_model(
  net_file,
  input_data_1,
  reference_samples,
  input_data_2,
  maxCat
)
}
\arguments{
\item{net_file}{Path to TXT file containing the reference sample network}

\item{input_data_1}{Dataset for LaGrACE model training}

\item{reference_samples}{Logical vector indicating which samples belong to
reference group (TRUE) and which do not (FALSE)}

\item{input_data_2}{Dataset for LaGrACE features calculation}

\item{maxCat}{Maximum number of categories for a variable to be treated as discrete}
}
\value{
A list containing a dataframe of the LaGrACE features, a dataframe
indicating which variables were selected for the prediction in calculating each feature
, and a dataframe including continuous/discrete feature information
}
\description{
This function train LaGrACE predictive model by a given reference dataset
and calculate LaGrACE featur (deviation) for another dataset
}
