% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_LaGrACE_deviation.R
\name{Calculate_LaGrACE_deviation}
\alias{Calculate_LaGrACE_deviation}
\title{Scan over a range of penalty value and calculate k-fold LaGrACE deviation for a given dataset.}
\usage{
Calculate_LaGrACE_deviation(input_data, pd_value, maxCat = 5, out_dir)
}
\arguments{
\item{input_data}{Dataframe containing data for all samples (rows) and variables (columns)}

\item{pd_value}{Penalty discount value for learning FGES causal network}

\item{maxCat}{Maximum number of categories for a variable to be treated as discrete}

\item{out_dir}{Path to directory to which all output files should be written}
}
\value{
A list containing a dataframe of the LaGrACE deviation value (absolute average deviation)
and causal graph information (#edges, # connected variables) over a range of Penalty discount value
}
\description{
Calculates k-fold LaGrACE deviation over a range of Penalty discount value for optimal penalty
discount value selection
}
