% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{CalFittedSigma}
\alias{CalFittedSigma}
\title{Calculate the fitted value of A_ICA_I^T for given Sigma and delta.}
\usage{
CalFittedSigma(Sigma, delta, Ms, Union, diagonal)
}
\arguments{
\item{Sigma}{p by p covariance matrix.}

\item{delta}{given parameter.}

\item{Ms}{the calculated maximal values of Sigma per row.}

\item{Union}{TRUE or FALSE.}

\item{diagonal}{TRUE or FALSE.}
}
\value{
a list including:
\itemize{
\item pureVec: vector of the indices of estimated pure variables.
\item fitted: fitted value of A_ICA_I^T
}
}
\description{
Calculate the fitted value of A_ICA_I^T for given Sigma and delta.
}
