% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{calFNRandFPR}
\alias{calFNRandFPR}
\title{Calculate the FNR and FPR for given p by K matrix A.}
\usage{
calFNRandFPR(estA, trueA)
}
\arguments{
\item{estA, trueA}{two matrices having same dims.}
}
\value{
a list including:
\itemize{
\item list of error: FNR and FPR.
\item errorInd: the row indices where errors occur.
}
}
\description{
Calculate the FNR and FPR for given p by K matrix A.
}
