% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{calSPandSN}
\alias{calSPandSN}
\title{Calculate specificity and sensitivity based on group partition}
\usage{
calSPandSN(p, estGroup, trueGroup)
}
\arguments{
\item{p}{the total number of variables}

\item{estGroup}{list of variable indices for estimated group}

\item{trueGroup}{list of variable indices for true group}
}
\value{
Specificity and Sensitivity
}
\description{
Calculate specificity and sensitivity based on group partition
}
