% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{CheckDisjoint}
\alias{CheckDisjoint}
\title{Check if all the sets are disjoint}
\usage{
CheckDisjoint(groupList)
}
\arguments{
\item{groupList}{groups of clustering}
}
\value{
TRUE or FALSE.
}
\description{
Check if all the sets are disjoint
}
