% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{checkElement}
\alias{checkElement}
\title{Check if an element is in a list. If it does exist in some group, return the group
index in that list and its sublist index. Otherwise, return c(0,0).}
\usage{
checkElement(element, groupList)
}
\arguments{
\item{element}{a variable}

\item{groupList}{list of variable indices for each group}
}
\value{
index of element in the list and its sublist, If the element is not in the list, return c(0,0)
}
\description{
Check if an element is in a list. If it does exist in some group, return the group
index in that list and its sublist index. Otherwise, return c(0,0).
}
