% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sample.R
\name{cluster_sample}
\alias{cluster_sample}
\title{Cluster samples based on LaGrACE feature}
\usage{
cluster_sample(
  features,
  meta_data_df,
  fgs_pd,
  n_pc,
  total_PC = 50,
  cluster_resolution,
  out_dir,
  cluster_ref_samples = FALSE,
  ref_samples,
  seed_value = 1,
  Label = NULL
)
}
\arguments{
\item{features}{LaGrACE features matrix/dataframe}

\item{meta_data_df}{Dataframe containing metadata for all samples}

\item{fgs_pd}{Sparsity Penalty Value for FGES algorithm}

\item{n_pc}{number of principal components  for clustering and UMAP embedding}

\item{total_PC}{total number of principal components for clustering and UMAP embedding}

\item{cluster_resolution}{Value of the resolution parameter, use a value above (below) 1.0 if
you want to obtain a larger (smaller) number of communities.}

\item{out_dir}{output directory for clustering information and visualization (Dimensional Reduction Plot: UMAP)}

\item{cluster_ref_samples}{Logical vector (TRUE/FALSE) indicating whether reference samples should be excluded/kept
from sample clustering}

\item{ref_samples}{Logical vector indicating which samples belong to
reference group (TRUE) and which do not (FALSE)}

\item{seed_value}{Seed of the random number generator}

\item{Label}{a string including label information for Dimensional Reduction Plot(UMAP)}
}
\value{
list including one Dataframe containing clustering information and one Matrix containing PC loading information
}
\description{
This function utilize graph-based unsupervised clustering algorithm (Seurat) for sample clustring
}
