% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_tree.R
\name{clustering_tree}
\alias{clustering_tree}
\title{Plotting clustering tree}
\usage{
clustering_tree(
  seurat_obj,
  n_pc = 10,
  cluster_resolution = seq(0, 1, 0.1),
  out_dir = "./",
  seed_value = 1
)
}
\arguments{
\item{seurat_obj}{Seurat Ojbect}

\item{n_pc}{number of principal components (total 30 PCs) for clustering}

\item{cluster_resolution}{scan over cluster resolution (from 0 to 1)}

\item{out_dir}{output directory for clustering tree plot}

\item{seed_value}{Seed of the random number generator}
}
\description{
This function creates a clustering tree showing the relationship between clusterings at different resolutions.
}
