% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert_Catogorical_to_Binary.R
\name{Convert_Catogorical_to_Binary}
\alias{Convert_Catogorical_to_Binary}
\title{Convert discrete variable into binary indicator variables in a dataframe}
\usage{
Convert_Catogorical_to_Binary(data, maxCat = 5)
}
\arguments{
\item{data}{a matrix or dataframe, each row represents a sample, each column represents a feature}

\item{maxCat}{Full dataset of with all samples}
}
\value{
A list containing a dataframe of the transformed dataset, a dataframe
indicating which variables are continuous, and a dataframe including
original variable names of transformed data
, and a dataframe including continuous/discrete feature information
}
\description{
This function transform all discrete variables into several binary indicator variables,
one for each category of the original discrete variable
}
