% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{CV_Delta}
\alias{CV_Delta}
\title{Cross validation for choosing delta. For each delta from the given grids, first split the data into two datasets.
Obtain I, AI and C. Then calculate AICAI^T and choose delta which minimizes the criterion: AICAI^t - Sigma(dataset 2)}
\usage{
CV_Delta(X, deltaGrids, Union, diagonal)
}
\arguments{
\item{X}{n by p matrix.}

\item{deltaGrids}{vector of numerical constants.}

\item{Union}{TRUE or FALSE for estimating the pure variables. (For details, see doc of FindPureNode.)}

\item{diagonal}{TRUE or FALSE for the diagonal structure of C.}
}
\value{
the selected optimal delta
}
\description{
Cross validation for choosing delta. For each delta from the given grids, first split the data into two datasets.
Obtain I, AI and C. Then calculate AICAI^T and choose delta which minimizes the criterion: AICAI^t - Sigma(dataset 2)
}
