% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{CV_lbd}
\alias{CV_lbd}
\title{Use Cross-validation to select lambda for estimating Omega.}
\usage{
CV_lbd(X, lbdGrids, AI, pureVec, diagonal)
}
\arguments{
\item{X}{n by p matrix.}

\item{lbdGrids}{vector of numerical constants.}

\item{AI}{p by K matrix}

\item{pureVec}{vector of indices of pure variables.}

\item{diagonal}{TRUE or FALSE for the diagonal structure of C.}
}
\value{
the selected optimal lambda
}
\description{
Split the data into two parts. Estimating C on two datasets. Then, for each lambda, calculate Omega on the first dataset and calculate the loss on the second dataset.
Find the lambda which gives the smallest loss of <C,Omega> - log(det(Omega))
}
