% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{EstAI}
\alias{EstAI}
\title{Use the given {@code delta} to calculate the fitted AI, pure variable indices in list form and vector form. Also return estimated Y and C for the following Dantzig estimation.}
\usage{
EstAI(Sigma, delta, Union)
}
\arguments{
\item{Sigma}{p by p covariance matrix.}

\item{delta}{delta to be used.}

\item{Union}{flag for using union or intersection for p sets of pure nodes.}
}
\value{
the list including:
\itemize{
\item AI: the p by K estimated AI
\item pureVec: vector of the indices of estimated pure variables
\item pureSignInd: list of the indices of estimated pure variables
}
}
\description{
Use the given {@code delta} to calculate the fitted AI, pure variable indices in list form and vector form. Also return estimated Y and C for the following Dantzig estimation.
}
