% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{EstAJInv}
\alias{EstAJInv}
\title{This function estimates the |J| by K matrix A_J by using soft thresholding.}
\usage{
EstAJInv(Omega, Y, lbd)
}
\arguments{
\item{Omega}{K by K estimated C^{-1}.}

\item{Y}{K by |J| reponse matrix.}

\item{lbd}{the chosen constant of the RHS constraint for soft-thresholding.}
}
\value{
|J| by K matrix A_J.
}
\description{
This function estimates the |J| by K matrix A_J by using soft thresholding.
}
