% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{EstC}
\alias{EstC}
\title{Estimate C. If diagonal=True, estimate only diagonal elements.}
\usage{
EstC(Sigma, AI, diagonal)
}
\arguments{
\item{Sigma}{p by p covariance matrix.}

\item{AI}{p by K matrix.}

\item{diagonal}{TRUE or FALSE.}
}
\value{
K by K estimated C_hat
}
\description{
Estimate C. If diagonal=True, estimate only diagonal elements.
}
