% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{EstOmega}
\alias{EstOmega}
\title{This function estimate omega}
\usage{
EstOmega(lbd, C)
}
\arguments{
\item{lbd}{list of lambda values}

\item{C}{K by K estimated C_hat}
}
\value{
omega
}
\description{
For a given lbd and C, solve the C^{-1}.
Require: C should be symmetric and square.
}
