% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{EstY}
\alias{EstY}
\title{Estimate the Sigma_{TJ}_hat}
\usage{
EstY(Sigma, AI, pureVec)
}
\arguments{
\item{Sigma}{p by p matrix.}

\item{AI}{p by K matrix.}

\item{pureVec}{the vector of pure node indices.}
}
\value{
Sigma_{TJ}_hat K by |J| matrix.
}
\description{
Estimate the Sigma_{TJ}_hat
}
