% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_variance.R
\name{filter_by_variance}
\alias{filter_by_variance}
\title{Filter a dataframe based on the variance of the columns.}
\usage{
filter_by_variance(df, n_var)
}
\arguments{
\item{df}{Dataframe containing data for all samples (rows) and variables (columns)}

\item{n_var}{Number of variables to keep for variance filter}
}
\value{
Dataframe containing the data filtered for highest column variance
}
\description{
This function filters a dataframe for the columns with the highest variance.
}
