% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markov_blanket_Pattern.R
\name{find_markov_blanket_Pattern}
\alias{find_markov_blanket_Pattern}
\title{Find the Markov blanket around a target variable.}
\usage{
find_markov_blanket_Pattern(target, graph)
}
\arguments{
\item{target}{A string containing the name of the target vertex}

\item{graph}{A list containing graph information}
}
\value{
A list of the names of vertices in the Markov blanket
}
\description{
This function returns the Markov blanket around a target vertex from a
graph
}
