% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{FindPureNode}
\alias{FindPureNode}
\title{Estimate list of pure node indices for given {@code Sigma} and {@delta}.}
\usage{
FindPureNode(Sigma, delta, Ms, Union)
}
\arguments{
\item{Sigma}{p by p sample covariance matrix.}

\item{delta}{numerical constant.}

\item{Ms}{the largest absolute values of each row of Sigma.}

\item{Union}{flag for using union or intersection for p sets of pure nodes.}
}
\value{
a list including:
\itemize{
\item the list of the estimated pure node indices
\item the vector of the estimated pure node indices
}
}
\description{
Estimate list of pure node indices for given {@code Sigma} and {@delta}.
}
