% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{FindRowMax}
\alias{FindRowMax}
\title{Calculate the maximal absolute value for each row of the given matrix.}
\usage{
FindRowMax(Sigma)
}
\arguments{
\item{Sigma}{p by p matrix}
}
\value{
length p vector
}
\description{
Calculate the maximal absolute value for each row of the given matrix.
}
