% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{FindSignPureNode}
\alias{FindSignPureNode}
\title{Estimate the sign subpartition of pure node sets. If there is an element of a list is empty, then a empty list will be put in that position}
\usage{
FindSignPureNode(pureList, Sigma)
}
\arguments{
\item{pureList}{list of pure node indices (Example: list(c(1,2,3),c(4,5,6,7)))}

\item{Sigma}{p by p sample covariance}
}
\value{
list of sign subpartition of pure node indices.
Example: list(list(c(1,2),3),list(c(4,7),c(5,6)))
}
\description{
Estimate the sign subpartition of pure node sets. If there is an element of a list is empty, then a empty list will be put in that position
}
