% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{genConstr}
\alias{genConstr}
\title{Obtain the Amat and bvec for a given C and lbd.}
\usage{
genConstr(lbd, C)
}
\arguments{
\item{lbd}{lamdba value}

\item{C}{K by K estimated C_hat}
}
\value{
a list including:
\itemize{
\item Amat
\item bvec
}
}
\description{
Obtain the Amat and bvec for a given C and lbd.
}
