% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_seuratobj.R
\name{generate_seuratobj}
\alias{generate_seuratobj}
\title{Generate Seurat Object}
\usage{
generate_seuratobj(
  features,
  meta_data_df,
  cluster_ref_samples = FALSE,
  ref_samples,
  Label = NULL
)
}
\arguments{
\item{features}{LaGrACE features matrix/dataframe}

\item{meta_data_df}{Dataframe containing metadata for all samples}

\item{cluster_ref_samples}{Logical vector (TRUE/FALSE) indicating whether reference samples should be excluded/kept
from sample clustering}

\item{ref_samples}{Logical vector indicating which samples belong to
reference group (TRUE) and which do not (FALSE)}

\item{Label}{a string including label information for Dimensional Reduction Plot(UMAP)}
}
\value{
a Seurat object
}
\description{
This function creates a Seurat object for LaGrACE feature
}
