% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{getErrors}
\alias{getErrors}
\title{Infer error between estimated and true matrices}
\usage{
getErrors(estA, trueA, flagPerm = TRUE)
}
\arguments{
\item{estA}{estimated matrix}

\item{trueA}{true matrix}

\item{flagPerm}{permutation or not}
}
\value{
a list of (specificity and sensitivity, error,matrix sup norm, frobenius norm(scaled) and L_inf norm)
}
\description{
Infer error between estimated and true matrices
}
