% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{getOptPerm}
\alias{getOptPerm}
\title{For the given pure node set, find the best column permutation matrix
Aperm of A such that |Aperm - B|_F is minimal for the target B}
\usage{
getOptPerm(A, B)
}
\arguments{
\item{A}{generic matrix.}

\item{B}{generic matrix with the same dimension of A.}
}
\value{
list of column permutation and sign permutation
}
\description{
For the given pure node set, find the best column permutation matrix
Aperm of A such that |Aperm - B|_F is minimal for the target B
}
