% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{lInfConstr}
\alias{lInfConstr}
\title{Get the Amat constraint matrix of Inf-1 norm, i.e. the l_1 norm of each row.}
\usage{
lInfConstr(p)
}
\arguments{
\item{p}{numerical constant denoting the dimension of Omega.}
}
\value{
p by p(p+1) matrix.
}
\description{
Get the Amat constraint matrix of Inf-1 norm, i.e. the l_1 norm of each row.
}
