% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{LOVE}
\alias{LOVE}
\title{Run Love Algorithm}
\usage{
LOVE(
  X,
  delta = seq(1.5, 2.5, 0.1),
  HT = F,
  lbd = 1,
  mu = 1,
  diagonal = FALSE,
  stand = TRUE,
  N = 50,
  Union = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{X}{n by p data matrix.}

\item{delta}{grid of leading constants for tuning delta(cluster anchor tuning). This grid needs to be multiplied by sqrt(log(p)/n)}

\item{HT}{If TRUE, then hard-thresholding is used for estimating A; else soft-thresholding is used}

\item{lbd}{membership matrix tuning}

\item{mu}{the grid of leading constants for tuning the thresholding level diagonal: TRUE if the cov(Z) is diagonal; else FALSE. If TRUE, OLS is used for solving A}

\item{diagonal}{TRUE if the cov(Z) is diagonal; else FALSE. If TRUE, #     OLS is used for solving A}

\item{stand}{True if centering is needed}

\item{N}{the number of cross validation used for delta}

\item{Union}{True if using "or" rule for estimating pure nodes; False if using "and" rule}

\item{verbose}{if the process needs to be printed out}
}
\value{
a list including:
\itemize{
\item K the number of clusters.
\item A the estimated p by K matrix.
\item group: the estimated group structure(indices for each cluster).
\item pure the vector of estimated pure variables.
}
}
\description{
Non-overlapping/Overlappig variable clustering for given data X where each row denotes a data point and columns for variables.
Reference: Bing X, Bunea F, Royer M, Das J. Latent Model-Based Clustering for Biological Discovery. iScience. 2019 Apr 26;14:125-135. doi: 10.1016/j.isci.2019.03.018. Epub 2019 Mar 21. PMID: 30954780; PMCID: PMC6449745.
}
