% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE_Delta_Tuning.R
\name{LOVE_Delta_Tuning}
\alias{LOVE_Delta_Tuning}
\title{Tuning Delta value for LOVE algorithm}
\usage{
LOVE_Delta_Tuning(Y_features, delta_seq)
}
\arguments{
\item{Y_features}{n by p data matrix.}

\item{delta_seq}{grid of constants for tuning delta(cluster anchor tuning)}
}
\value{
a list including:
\itemize{
\item delta_seq
\item AIC score
\item BIC score
\item LOVE latent factor
}
}
\description{
Tuning Delta value for LOVE algorithm
}
