% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{offSum}
\alias{offSum}
\title{Calculate the sum of squares of the upper off-diagonal elements of two matrices (require: M and N have the same dimensions)}
\usage{
offSum(M, N)
}
\arguments{
\item{M}{Matrix M}

\item{N}{Matrix N}
}
\value{
sum of squares of the upper off-diagonal elements of two matrices
}
\description{
Calculate the sum of squares of the upper off-diagonal elements of two matrices (require: M and N have the same dimensions)
}
