% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{permA}
\alias{permA}
\title{Find the optimal permutated matrix of A such that |Aperm-B|_F is minimized.}
\usage{
permA(A, B)
}
\arguments{
\item{A, B}{two matrices with the same dimension.}
}
\value{
Permutated A.
}
\description{
Find the optimal permutated matrix of A such that |Aperm-B|_F is minimized.
}
