% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{prodConstr}
\alias{prodConstr}
\title{Get the Amat constraint matrix for the product Omega \%\emph{\% C based on
diagonal and off-diagonal such that:
t(A_i) \%}\% par_vec <- t(Omega_i) \%*\% C_i,
for i belonging to the diagonal set. Similar for the off-diagonal,
the order for each row of off-diagonal set is from the first row to
the last row and ordered by row.}
\usage{
prodConstr(C)
}
\arguments{
\item{C}{K by K matrix.}
}
\value{
a list including:
\itemize{
\item the constraint matrices of diagonal and offdiagonal.
}
}
\description{
Get the Amat constraint matrix for the product Omega \%\emph{\% C based on
diagonal and off-diagonal such that:
t(A_i) \%}\% par_vec <- t(Omega_i) \%*\% C_i,
for i belonging to the diagonal set. Similar for the off-diagonal,
the order for each row of off-diagonal set is from the first row to
the last row and ordered by row.
}
