% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{pureRowInd}
\alias{pureRowInd}
\title{For given matrix A, find the row indices which correspond to pure nodes}
\usage{
pureRowInd(A)
}
\arguments{
\item{A}{p by K matrix.}
}
\value{
vector of pure row indices
}
\description{
For given matrix A, find the row indices which correspond to pure nodes
}
