% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_causality_graph.R
\name{read_causality_graph}
\alias{read_causality_graph}
\title{Load reference sample causal network}
\usage{
read_causality_graph(file)
}
\arguments{
\item{file}{Path to TXT file containing the reference sample network}
}
\value{
A list containing edge, nodes and adjacency information of causal graph learned by FGES
}
\description{
this functions loads reference sample causal network (FGES) from TXT file
}
