% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{RecoverAI}
\alias{RecoverAI}
\title{Recover the estimated submatrix A_I by given the pure node group.}
\usage{
RecoverAI(estGroupList, p)
}
\arguments{
\item{estGroupList}{list of group indices of the pure node with sign.}

\item{p}{number of rows}
}
\value{
p by K matrix.
}
\description{
Recover the estimated submatrix A_I by given the pure node group.
}
