% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{recoverGroup}
\alias{recoverGroup}
\title{Recover group structure based given p by K matrix A and perform thresholding}
\usage{
recoverGroup(A)
}
\arguments{
\item{A}{estimated matrix. thresh constant > 0.}
}
\value{
list of group indices with sign subpartition
}
\description{
Recover group structure based given p by K matrix A and perform thresholding
}
