% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LaGrACE.R
\name{run_LaGrACE}
\alias{run_LaGrACE}
\title{Learn FGES network and calculate LaGrACE divergence score}
\usage{
run_LaGrACE(input_data, reference_samples, fges_pd = 1, maxCat = 5, out_dir)
}
\arguments{
\item{input_data}{Dataframe containing data for all samples (rows) and variables (columns)}

\item{reference_samples}{Logical vector indicating which samples belong to
reference group (TRUE) and which do not (FALSE)}

\item{fges_pd}{Penalty discount value for learning FGES causal network}

\item{maxCat}{Maximum number of categories for a variable to be treated as discrete}

\item{out_dir}{Path to directory to which all output files should be written}
}
\value{
Path to file containing the output from FGES
}
\description{
This function returns LaGrACE divergence score
}
