% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{setConstr}
\alias{setConstr}
\title{For a given ith row of the matrix Omega and {@code vec}, find the vec_constr}
\usage{
setConstr(i, vec)
}
\arguments{
\item{i}{is an index of each row, 1 <= i <= n.}

\item{vec}{length 2*p vector containing the values to be modified.}
}
\value{
length p*(p+1) vector.
eg. |C21| + |C22| + |C23| = t(c(1,1,1,1,1,1) \%\emph{\% c(c21(+),c21(-),c22(+),c22(-),c23(+),c23(-))
= t(c(0,0,1,-1,0,0,1,-1,1,-1,0,0)) \%}\% par_vec
}
\description{
such that t(vec_constr) \%\emph{\% par_vec = 't(vec) \%}\% (ci1(+),ci1(-),ci2(+),ci2(-),ci3(+),ci3(-)).
}
