% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{signPerm}
\alias{signPerm}
\title{For a given integer K, generate its sign permutation}
\usage{
signPerm(n)
}
\arguments{
\item{n}{an integer K}
}
\value{
a 2^K by K matrix containing all sign permutations
}
\description{
For a given integer K, generate its sign permutation
}
