% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{singleton}
\alias{singleton}
\title{Check if there exists an element of the given list has length equal to 1}
\usage{
singleton(estPureIndices)
}
\arguments{
\item{estPureIndices}{list of pure indices}
}
\value{
True or False
}
\description{
If exists at least one, return TRUE; otherwise return FALSE
}
