% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{Solve_row}
\alias{Solve_row}
\title{This function solve LP program}
\usage{
Solve_row(col_ind, C, lbd)
}
\arguments{
\item{col_ind}{K by K estimated C^{-1}.}

\item{C}{K by |J| reponse matrix.}

\item{lbd}{the chosen constant of the RHS constraint for soft-thresholding.}
}
\value{
|J| by K matrix A_J.
}
\description{
This function solve LP program
}
