% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{TestPure}
\alias{TestPure}
\title{For given row, check if it is a pure node by interactively checking the nodes in Si. Return TRUE if the given row corresponds to a pure variable.}
\usage{
TestPure(Sigma, rowInd, Si, Ms, delta)
}
\arguments{
\item{Sigma}{p by p matrix.}

\item{rowInd}{integer index}

\item{Si}{vector of indices.}

\item{Ms}{vector of largest absolute values of each rows in Si}

\item{delta}{numerical constant.}
}
\value{
TRUE or FALSE.
}
\description{
For given row, check if it is a pure node by interactively checking the nodes in Si. Return TRUE if the given row corresponds to a pure variable.
}
