% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{threshA}
\alias{threshA}
\title{Normalize matrix based on threshold}
\usage{
threshA(A, mu, scale = FALSE)
}
\arguments{
\item{A}{estimated matrix}

\item{mu}{threshold}

\item{scale}{normalize or not}
}
\value{
normalized matrix
}
\description{
Threshold the estimated {@code A} based on the given {@code mu}. If {@code scale} is true,
then normalize each row of A such that the l-1 norm of each row is not larger than 1.
}
