% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOVE.R
\name{UnionGroup}
\alias{UnionGroup}
\title{Union all the groups which have common variables}
\usage{
UnionGroup(groupList, Union)
}
\arguments{
\item{groupList}{groups of clustering}

\item{Union}{flag for using union or intersection for p sets of pure nodes.}
}
\value{
TRUE or FALSE.
}
\description{
Union all the groups which have common variables
}
